Move resolvedTextEntryMode into AIChatSettings and update Package.resolved
